+++ test/bigtest/network/config.js
@@ -150,6 +150,10 @@ export default function configure() {
           inst.contributors[0].name.match(right.term));
       }
 
+      if (field === 'allTitles') {
+        return instances.all().filter(inst => inst.alternativeTitles.some(el => el.alternativeTitle === term));
+      }
+
       // With the addition of a third condition to search for records by 'Holdings. Call number readable by eye" in the CQL query,
       // cqlParser.tree object gets a deeper structure in the 'left' and 'right' properties.
       if (left?.left?.field === 'holdingsRecords.fullCallNumber') {